LayoutTests/streams/streams-public-array-api-expected.txt

Streams should not directly use array public APIs

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS result.value is "half baked potato"
PASS result.value is "fully baked potato"
PASS successfullyParsed is true

TEST COMPLETE
